Replaces edit and create modals Original modal was dropped, editSurvivorDialog was made multi-purpose
