+++ web/src/pages/settlement/population/tab.tsx
@@ -1,40 +1,46 @@
 import {useState} from 'react';
 import {SurvivorTable} from './survivorTable';
-import EditSurvivorDialog from './editSurvivorDialog';
-import {type Survivor} from '~/lib/survivor';
+import SurvivorDialog from './survivorDialog';
+import {type Survivor, SurvivorTemplate} from '~/lib/survivor';
 import {useLoaderData, useParams, useRevalidator} from 'react-router';
 import styles from './tab.module.css';
 
+type DialogState = Survivor | null
 
 export function PopulationTab() {
   const {settlementId} = useParams();
   const survivors = (useLoaderData() as Survivor[]) ?? [];
   const revalidator = useRevalidator();
-  const [editingSurvivor, setEditingSurvivor] = useState<Survivor | null>(null);
+  const [dialogState, setDialogState] = useState<DialogState>(null);
 
   if (!settlementId) {
     throw Error('settlement id is required');
   }
 
-  const handleEditSuccess = () => {
-    setEditingSurvivor(null);
+  const handleSuccess = () => {
+    setDialogState(null);
     revalidator.revalidate();
+  };
+
+  const handleClose = () => {
+    setDialogState(null);
   };
 
   return (
     <div id="population" className={styles.tab}>
       <SurvivorTable
         data={survivors}
-        settlementId={settlementId}
-        onEditSurvivor={setEditingSurvivor}
-        onSurvivorCreated={() => revalidator.revalidate()}
+        onEditSurvivor={(survivor) => setDialogState(survivor)}
+        onCreateSurvivor={() => setDialogState(SurvivorTemplate(settlementId))}
       />
-      <EditSurvivorDialog
-        survivor={editingSurvivor}
-        onClose={() => setEditingSurvivor(null)}
-        onSuccess={handleEditSuccess}
-      />
+      {dialogState && (
+        <SurvivorDialog
+          data={dialogState}
+          settlementId={settlementId}
+          onClose={handleClose}
+          onSuccess={handleSuccess}
+        />
+      )}
     </div>
   );
 }
-
